// src/interfaces/coin.interface.ts
export interface Icoin {
  id: string;
  symbol: string;
  name: string;
  nameid: string;
  rank: number;
  price_usd: String;
  percent_change_24h: String;
  percent_change_1h: number;
  percent_change_7d: String;
  price_btc: String;
  market_cap_usd: String;
  volume24: Number;
  volume24a: Number;
  csupply: String;
  tsupply: String;
  msupply: String;
}